KPERET-11743: Add plugin to simplemde-markdown-editor
- Expand our JSX plugin so we can now support not self closing components like: <ButtonA ...props></ButtonA>

This is because for DynamicModalLinkA the backend is not able tos end us the component like <DynamicModalLinkA ...props /> due to a limitation with CodeMirror and its replaceSelection method that automatically adds the close tag: </DynamicModalLinkA>.

Keep in mind that children needs to be part of the explicit prop "children" for the component just like we are doing it right now, i.e, do not expect to place content between the tags and parse those properly. For Example: <DynamicModalLinkA>Hello World</DynamicModalLinkA> won't work.